Fix login test
